backend/requirements.txt
annotated-types==0.7.0
anthropic==0.85.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
distro==1.9.0
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
docstring_parser==0.17.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-texttospeech==2.34.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
proto-plus==1.27.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.5
stripe==14.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
